Conflicts:
	deps/tracemonkey/jsscope.h
	deps/tracemonkey/jsstr.h
	deps/tracemonkey/jstypedarray.cpp
	include/nglKernel.h
	include/nuiMainWindow.h
	include/nuiRefCount.h
	nui3.xcodeproj/project.pbxproj
	src/Application/Cocoa/nglApplication_Cocoa.mm
	src/Audio/AudioUnit/nuiAudioDevice_AudioUnit.mm
	src/Audio/CoreAudio/nuiAudioDevice_CoreAudio.cpp
	src/Decorations/nuiDecoration.cpp
	src/String/nglString.cpp
	src/WidgetTree/nuiTopLevel.cpp
	src/Window/UIKit/nglWindow_UIKit.mm

+++ src/Image/nglImageCGCodec.cpp
@@ -18,14 +18,9 @@
 */
 
 #include "nui.h"
-
 #include NGL_CONFIG_H
-
-#include "nglImage.h"
 #include "nglImageCGCodec.h"
 
-#include "nglIStream.h"
-#include "nglOStream.h"
 
 #if (defined _COCOA_) || (defined _CARBON_)
 #import <ApplicationServices/ApplicationServices.h>
@@ -228,7 +223,7 @@ bool nglImageCGCodec::Feed(nglIStream* pIStream)
                           pCGColors, kCGImageAlphaPremultipliedLast);//kCGImageAlphaPremultipliedFirst);
 
   CGColorSpaceRelease(pCGColors);
-  CGRect rect = { {0,0}, {info.mWidth, info.mHeight} };
+  CGRect rect = { {0,0}, {static_cast<CGFloat>(info.mWidth), static_cast<CGFloat>(info.mHeight)} };
   CGContextClearRect(pCGContext, rect);
   CGContextDrawImage(pCGContext, rect, mpCGImage);
   CGContextRelease(pCGContext);
